refactor fragments destination
