+++ app/src/main/java/com/example/shift/chat/ChatFragment.kt
@@ -1,4 +1,4 @@
-package com.example.shift
+package com.example.shift.chat
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
