+++ app/src/main/java/com/example/shift/wallet/WalletFragment.kt
@@ -1,20 +1,20 @@
-package com.example.shift
+package com.example.shift.wallet
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.shift.databinding.FragmentRentHistoryBinding
+import com.example.shift.databinding.FragmentWalletBinding
 
-class RentHistoryFragment : Fragment() {
-    private lateinit var binding: FragmentRentHistoryBinding
+class WalletFragment : Fragment() {
+    private lateinit var binding: FragmentWalletBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentRentHistoryBinding.inflate(inflater, container, false)
+        binding = FragmentWalletBinding.inflate(inflater, container, false)
 
         return binding.root
     }
