+++ app/src/main/res/layout/fragment_chat.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:soundEffectsEnabled="false"
-    tools:context=".ChatFragment">
+    tools:context=".chat.ChatFragment">
 
     <TextView
         android:id="@+id/toolsTextView"
